method_2_gemini/requirements.txt
google-genai
python-dotenv
pyautogui
mss
Pillow
requests
